Added categories seeder
