+++ VictoryCenter/VictoryCenter.IntegrationTests/Utils/VictoryCenterWebApplicationFactory.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using VictoryCenter.DAL.Data;
 using VictoryCenter.DAL.Entities;
+using VictoryCenter.IntegrationTests.Utils.Seeder;
 
 namespace VictoryCenter.IntegrationTests.Utils;
 
@@ -29,8 +30,8 @@ public class VictoryCenterWebApplicationFactory<T> : WebApplicationFactory<T> wh
             var dbContext = scope.ServiceProvider.GetRequiredService<VictoryCenterDbContext>();
             dbContext.Database.EnsureCreated();
             
-            TestDatabaseSeeder.DeleteExistingData(dbContext);
-            TestDatabaseSeeder.SeedData(dbContext);
+            IntegrationTestsDatabaseSeeder.DeleteExistingData(dbContext);
+            IntegrationTestsDatabaseSeeder.SeedData(dbContext);
         });
     }
